Extract deployment code
We extract the deployment code that was split between
`src/bin/osc-deploy.rs` and `src/deploy.rs` into its own crate. This
allows us to reuse this code for more involved testing.
