+++ .gitignore
@@ -1,3 +1,4 @@
+/*/target
 /target
 **/*.rs.bk
 .oscoin_ledger_address
