+++ deploy/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "oscoin_deploy"
+description = "Tools to deploy the oscoin ledger to a node"
+version = "0.1.0"
+authors = ["Thomas Scholtes <thomas@monadic.xyz>"]
+edition = "2018"
+
+[dependencies]
+web3 = { git = "https://github.com/tomusdrw/rust-web3.git", rev = "b2aa7336bc9bf192f7959e79525a6d2667ff4c85" }
+env_logger = "0.6.2"
+hex = "0.3.1"
+clap = "2.31"
